Refactor to show table only on desktop
