+++ src/components/Movies/List/List.jsx
@@ -1,0 +1,7 @@
+import "./List.css";
+
+const List = ({ moviesArray, fetchNextPage }) => {
+  return <div>List</div>;
+};
+
+export default List;
