+++ src/components/Movies/Table/Table.css
@@ -17,6 +17,7 @@
   font-family: "Roboto", sans-serif;
   min-width: 400px;
 
+  margin: 25px 0;
   border-radius: 5px 5px 0 0;
   overflow: hidden;
   box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
@@ -35,7 +36,7 @@
 
 .movie-table th,
 .movie-table td {
-  padding: 8px 10px;
+  padding: 12px 15px;
   border: 1px solid black;
 }
 .table-footer {
@@ -53,23 +54,6 @@
   outline: 4px auto -webkit-focus-ring-color;
 }
 .movie-table .movie-image {
-  height: 115px;
+  height: 155px;
   aspect-ratio: 2/3;
-}
-
-@media (min-width: 768px) {
-  .movie-table--outer-container {
-    height: 467px;
-  }
-  .movie-table {
-    margin: 25px 0;
-    width: 100%;
-  }
-  .movie-table th,
-  .movie-table td {
-    padding: 12px 15px;
-  }
-  .movie-table .movie-image {
-    height: 155px;
-  }
 }
