+++ src/components/StarRating/Star/Star.jsx
